Use dir header first_free_byte as additional termination
Directory entries at the end of the block are supposed to be flagged as such.
However, some images built with 3doiso.exe appear to not set the appropriate
flag indicating the end of a block. Need to additionally use the first_free_byte
value from the directory header and exit once that relative offset is hit.